+++ example/redux/reducer.tsx
@@ -1,5 +1,0 @@
-import { sampleSlice } from 'redux/slice/sample'
-
-export const reducer = {
-  sample: sampleSlice.reducer,
-}