+++ example/types/sample.tsx
@@ -1,6 +1,0 @@
-export interface SampleType {
-  value1: string
-  value2: string
-}
-
-export default SampleType