+++ example/types/wallet.tsx
@@ -1,5 +1,0 @@
-export interface WalletType {
-
-}
-
-export default WalletType