+++ src/printer.ts
@@ -14,7 +14,7 @@ import { inject } from './generators/inject'
 export const Printer = new Command('🖨️ Printer')
 
 Printer
-  .version('1.2.0')
+  .version('1.2.1')
   .description('🖨️ Printer: Automation Tooling for Next, Redux and Prisma.')
   .option('-a, --no-action', 'do not inject actions', false)
   .option('-s, --no-state', 'do not inject state', false)